add workflow for continuous integration
Adds a workflow file for checking out the repository, installing python 3.8, installing tox and running the tests, to be run anytime that there is a push to the main branch of the repository.

The ambassador that will depend on this firewalling solution will require compatibility with Windows 7 and python 3.8 is the latest version which is compatible. To ensure compatibility with Windows 7 for this specifically, we are ensuring compatibility with python 3.8 and using that for our tests.